Fix warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:818:41: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
  818 |     memcpy(data, frame.get_data(), size * 3);
      |     ~~~~~~                         ~~~~~^~~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:823:41: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
  823 |     memcpy(data, frame.get_data(), size * 4);
      |     ~~~~~~                         ~~~~~^~~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:828:41: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
  828 |     memcpy(data, frame.get_data(), size * 2);
      |     ~~~~~~                         ~~~~~^~~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:832:36: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
  832 |     memcpy(data, frame.get_data(), size);
      |     ~~~~~~                         ^~~~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1121:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1121 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1121:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1121 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
      |                            ^ ~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1123:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1123 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1123:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1123 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
      |                            ^ ~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1125:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1125 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1125:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1125 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
      |                            ^ ~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1129:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1129 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1129:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1129 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
      |                            ^ ~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1131:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1131 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1131:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1131 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
      |                            ^ ~
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1133:73: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1133 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
      |                                                                       ~ ^
visp/modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp:1133:28: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
 1133 |             p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
      |                            ^ ~

+++ modules/sensor/src/rgb-depth/realsense/vpRealSense2.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2025 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -810,7 +810,7 @@ void vpRealSense2::getGreyFrame(const rs2::frame &frame, vpImage<unsigned char> 
 void vpRealSense2::getNativeFrameData(const rs2::frame &frame, unsigned char *const data)
 {
   auto vf = frame.as<rs2::video_frame>();
-  int size = vf.get_width() * vf.get_height();
+  size_t size = static_cast<size_t>(vf.get_width()) * static_cast<size_t>(vf.get_height());
 
   switch (frame.get_profile().format()) {
   case RS2_FORMAT_RGB8:
@@ -1118,19 +1118,19 @@ void vpRealSense2::getPointcloud(const rs2::depth_frame &depth_frame, const rs2:
 #else
         if (swap_rb) {
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].b =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel];
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].g =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel + 1];
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].r =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel + 2];
         }
         else {
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].r =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel];
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].g =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 1];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel + 1];
           pointcloud->points[static_cast<size_t>(depth_pixel_index)].b =
-            p_color_frame[(i * static_cast<unsigned int>(color_width) + j) * nb_color_pixel + 2];
+            p_color_frame[(static_cast<unsigned int>(i) * static_cast<unsigned int>(color_width) + static_cast<unsigned int>(j)) * nb_color_pixel + 2];
         }
 #endif
       }
